refactor(tx): split engine.go into focused files (#79)
Move the transaction engine's pipeline phases into separate files within
the same package. Pure code motion — no semantic changes, all behaviour
identical.

  engine.go        — Engine type, EngineConfig, LedgerView, NewEngine,
                     hash/owner-count helpers (302 lines)
  apply.go         — Apply, ApplyPseudo, applyPseudoTransaction
  preflight.go     — preflight + validation helpers
  preclaim.go      — preclaim + batch sign checks + signer lookup uses
  do_apply.go      — doApply with full tec/invariant recovery
  fee.go           — calculateFee, parseTxDeclaredFee
  reserve.go       — AccountReserve, ReserveForNewObject, etc.
  metadata.go      — ApplyResult, Metadata, AffectedNode JSON shaping
  signer_lookup.go — engineSignerListLookup helper
