+++ internal/tx/signer_lookup.go
@@ -1,0 +1,73 @@
+package tx
+
+import (
+	"errors"
+
+	"github.com/LeJamon/goXRPLd/internal/ledger/state"
+	"github.com/LeJamon/goXRPLd/keylet"
+)
+
+// engineSignerListLookup implements SignerListLookup using the engine's ledger view
+type engineSignerListLookup struct {
+	view LedgerView
+}
+
+// GetSignerList returns the signer list for an account
+func (l *engineSignerListLookup) GetSignerList(account string) (*state.SignerListInfo, error) {
+	accountID, err := state.DecodeAccountID(account)
+	if err != nil {
+		return nil, err
+	}
+
+	// Look up the signer list (SignerListID is always 0 currently)
+	signerListKey := keylet.SignerList(accountID)
+	exists, err := l.view.Exists(signerListKey)
+	if err != nil {
+		return nil, err
+	}
+	if !exists {
+		return nil, nil // No signer list
+	}
+
+	// Read and parse the signer list
+	signerListData, err := l.view.Read(signerListKey)
+	if err != nil {
+		return nil, err
+	}
+
+	signerList, err := state.ParseSignerList(signerListData)
+	if err != nil {
+		return nil, err
+	}
+
+	return signerList, nil
+}
+
+// GetAccountInfo returns account information needed for signer validation
+func (l *engineSignerListLookup) GetAccountInfo(account string) (flags uint32, regularKey string, err error) {
+	accountID, err := state.DecodeAccountID(account)
+	if err != nil {
+		return 0, "", err
+	}
+
+	accountKey := keylet.Account(accountID)
+	exists, err := l.view.Exists(accountKey)
+	if err != nil {
+		return 0, "", err
+	}
+	if !exists {
+		return 0, "", errors.New("account not found")
+	}
+
+	accountData, err := l.view.Read(accountKey)
+	if err != nil {
+		return 0, "", err
+	}
+
+	accountRoot, err := state.ParseAccountRoot(accountData)
+	if err != nil {
+		return 0, "", err
+	}
+
+	return accountRoot.Flags, accountRoot.RegularKey, nil
+}
